hack in changes to run tests against previously populated database

+++ api_tests/prepare-drone-federation-test.sh
@@ -3,15 +3,21 @@
 #   it is expected that this script is called by run-federation-test.sh script.
 set -e
 
+bypass_pg_purge=true;
+
 export RUST_BACKTRACE=1
 export RUST_LOG="warn,lemmy_server=debug,lemmy_api=debug,lemmy_api_common=debug,lemmy_api_crud=debug,lemmy_apub=debug,lemmy_db_schema=debug,lemmy_db_views=debug,lemmy_db_views_actor=debug,lemmy_db_views_moderator=debug,lemmy_routes=debug,lemmy_utils=debug,lemmy_websocket=debug"
 
+if [ "$bypass_pg_purge" = false ]
+then
+echo "ready PosgreSQL for drone instances"
 for INSTANCE in lemmy_alpha lemmy_beta lemmy_gamma lemmy_delta lemmy_epsilon; do
   echo "DB URL: ${LEMMY_DATABASE_URL} INSTANCE: $INSTANCE"
   psql "${LEMMY_DATABASE_URL}/lemmy" -c "DROP DATABASE IF EXISTS $INSTANCE"
   echo "create database"
   psql "${LEMMY_DATABASE_URL}/lemmy" -c "CREATE DATABASE $INSTANCE"
 done
+fi
 
 if [ -z "$DO_WRITE_HOSTS_FILE" ]; then
   if ! grep -q lemmy-alpha /etc/hosts; then
@@ -31,7 +37,7 @@ else
 fi
 
 echo "killall existing lemmy_server processes"
-killall -s1 lemmy_server || true
+killall lemmy_server || true
 
 echo "$PWD"
